requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
einops>=0.3.0
numpy
opencv-python 